+++ packages/images/optimize.d.ts
@@ -1,1 +1,1 @@
-export function svgOptions(): object;
+export function svgOptions(): Record<string, unknown>;